More Styling

+++ user_auth/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth import login , authenticate ,logout ,update_session_aut
 from django.contrib import messages
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
+from .models import *
 
 # imports for generate_password function
 import random
@@ -10,9 +11,12 @@ import random
 
 
 def home(request) :
+    user = get_object_or_404(User, pk=request.user.id)
+    userProfile = get_object_or_404(UserProfile, user=user)
     context = {
-        'title' : 'Home'
-
+        'title' : 'Home',
+        'user_': user,
+        'user_profile_': userProfile,
     }
     return render(request,'user_auth/home.html',context)
 
@@ -89,6 +93,9 @@ def sign_up_user(request) :
                                                 password = user_password1
                                                 )
                 user.save()
+                UserProfile.objects.create(
+                    user=user,
+                )
                 login(request,user)
                 return redirect("home")
 
@@ -98,7 +105,7 @@ def delete_account(request) :
     user = get_object_or_404(User , pk=request.user.id)
     if request.method == "POST" :
         text_confirm  = user.username+"-delete my account"
-        if request.POST['confirmation'] == text_confirm :
+        if (request.POST['confirmation'] == text_confirm) and (user.check_password(request.POST['password_for_deleting_account'])) :
             logout(request)
             user.delete()
             messages.info(request,"Your account was deleted sucssfully")
@@ -137,35 +144,42 @@ def edit_information(request) :
         }
         return render(request,"user_auth/home.html",context)
     elif request.method == "POST" :
-        user = get_object_or_404(User,pk=request.user.id)
+        user = get_object_or_404(User, pk=request.user.id)
+        userProfile = UserProfile.objects.get(user=user)
+
         user_first_name = request.POST['first_name']
         user_last_name = request.POST['last_name']
         user_username = request.POST['username']
         user_email = request.POST['email']
+        user_gender = request.POST['gender']
+
+        user.first_name = user_first_name
+        user.last_name = user_last_name
+        userProfile.gender = user_gender
+        userProfile.save()
+
         uniqe_username = User.objects.filter(username=user_username)
         uniqe_email = User.objects.filter(email=user_email)
         same_username = user.username == user_username
         same_email = user.email == user_email
+
+
+
         if (uniqe_username.__len__() == 0) and (uniqe_email.__len__() == 0) :
-            user.first_name = user_first_name
-            user.last_name = user_last_name
             user.username = user_username
             user.email = user_email
             user.save()
             messages.info(request,"the changes done sucssfully")
 
         elif (uniqe_username.__len__() == 0) and (uniqe_email.__len__() != 0) :
-            user.first_name = user_first_name
-            user.last_name = user_last_name
             user.username = user_username
             user.save()
             if not same_email :
                 messages.info(request,"the changes done , but the email is alrady taken , try another one !")
             else :
                 messages.info(request, "the changes done sucssfully")
+
         elif (uniqe_username.__len__() != 0) and (uniqe_email.__len__() == 0) :
-            user.first_name = user_first_name
-            user.last_name = user_last_name
             user.email = user_email
             user.save()
             if not same_username :
@@ -174,8 +188,6 @@ def edit_information(request) :
                 messages.info(request, "the changes done sucssfully")
 
         else :
-            user.first_name = user_first_name
-            user.last_name = user_last_name
             user.save()
             if (not same_username) and (not same_email) :
                 messages.info(request,"the changes done , but the username and email is alrady taken , try another ones!")
